% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acomoda_text_lavaan.R
\name{acomoda_text_lavaan}
\alias{acomoda_text_lavaan}
\title{Acomoda la sintaxis para el cfa}
\usage{
acomoda_text_lavaan(data_preg)
}
\arguments{
\item{data_preg}{Tabla con informacion del modelo a estimar. Filtramos la escala en el MIAU y aplicamos la funcion.}
}
\value{
Sintaxis para modelos lavaan::cfa
}
\description{
Toma la informacion del MIAU y acomoda para a la sintaxis que aplica lavaan::cfa
}
\examples{
\dontrun{
acomoda_text_lavaan(preg)
}


}
