% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_icc.R
\name{calc_icc}
\alias{calc_icc}
\title{Calcula el ICC}
\usage{
calc_icc(model_lmer)
}
\arguments{
\item{model_lmer}{Output de lme4::lmer en el modelo nulo}
}
\value{
icc del modelo nulo
}
\description{
Calcula el coeficiente de correlacion intraclase de un modelo nulo (ICC)
}
\examples{
\dontrun{
fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
calc_icc(fm1)
}

}
