% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_recursivo_lavaan.R
\name{cfa_recursivo_lavaan}
\alias{cfa_recursivo_lavaan}
\title{Realiza analisis factorial confirmatorio buscando una solucion razonable}
\usage{
cfa_recursivo_lavaan(data, model_lavaan, recursivo = TRUE, puntajes = TRUE)
}
\arguments{
\item{data}{dataframe}

\item{model_lavaan}{sintaxis de lavaan (text)}

\item{recursivo}{TRUE si se quiere recursivo}

\item{puntajes}{TRUE en caso se requieran los puntajes}
}
\value{
lista con indicadores de ajuste, cargas factoriales y confiabilidad
}
\description{
Si recursivo = TRUE, ejecuta un CFA hasta encontrar una solucion razonable. CFI y TLI > 0.95; SRMR y RMSEA < 0.10.
}
\examples{
\dontrun{
HS.model <- ' visual  =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed   =~ x7 + x8 + x9 '
cfa_recursivo_lavaan(lavaan::HolzingerSwineford1939, HS.model, recursivo = FALSE)
}

}
