% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2.R
\name{cor2}
\alias{cor2}
\title{Matriz de correlaciones}
\usage{
cor2(data, r = 2, long = NULL, corr = NULL)
}
\arguments{
\item{data}{data frame}

\item{r}{redondear valores}

\item{long}{en caso se requiera en formato dataframe}

\item{corr}{si colocas 'poly' aplica correlacion policorica}
}
\value{
matriz de correlaciones
}
\description{
Lo mismo que cor() pero con algunas acomodaciones. Si hay NA usa "pairwise.complete.obs".
}
\examples{
cor2(iris[1:4])

}
