% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_sig.R
\name{dif_sig}
\alias{dif_sig}
\title{Contraste de hipotesis}
\usage{
dif_sig(data, m, se, nom = NULL, long = NULL)
}
\arguments{
\item{data}{data frame}

\item{m}{estimacion}

\item{se}{error estandar de la estimacion}

\item{nom}{grupos}

\item{long}{en caso se quiera en long}
}
\value{
data frame indicando diferencias estadisticas
}
\description{
Toma la estimacion puntual y los errores estandar asociados para establecer diferencias estadisticas
}
\examples{
tabla1 <- data.frame(
 sex = c("Hombre", "Mujer"),
 mean1 = c(317.68, 335.34),
 se1 = c(7.22, 6.17)
 )
dif_sig(tabla1, m = "mean1", se = "se1", long = TRUE)


}
