% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_patron_missing.R
\name{g_patron_missing}
\alias{g_patron_missing}
\title{Visualizar patron de valores perdidos}
\usage{
g_patron_missing(data, cols = NULL, text.size = 5)
}
\arguments{
\item{data}{data frame}

\item{cols}{para especificar algunas columnas}

\item{text.size}{cambiar tamaño del texto de las columnas}
}
\value{
grafico
}
\description{
Visualizar patron de valores perdidos de un dataframe.
}
\examples{
g_patron_missing(data = iris)
g_patron_missing(data = iris, cols = c("Sepal.Length", "Sepal.Width"), text.size = 10)

}
