% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_prop_grupo.R
\name{mean_prop_grupo}
\alias{mean_prop_grupo}
\title{Promedio de una variable según grupos}
\usage{
mean_prop_grupo(data, m, w = NULL, vis_na = TRUE)
}
\arguments{
\item{data}{data frame}

\item{m}{variable continua}

\item{w}{en caso hayan pesos (if there are weights)}

\item{vis_na}{en caso se consideren valores perdidos}
}
\value{
data frame con el promedio según grupo, la cantidad de casos (n) y porcentaje (prop)
}
\description{
Promedio de una variable según grupos
}
\examples{
dplyr::group_by(iris, Species) |> mean_prop_grupo(m = "Sepal.Width", w = "Petal.Width")
}
