% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_hlm.R
\name{mplus_hlm}
\alias{mplus_hlm}
\title{Regresion multinivel en Mplus}
\usage{
mplus_hlm(data, y, x1 = NULL, x2 = NULL, peso_est, peso_ie, idie, r)
}
\arguments{
\item{data}{dataframe}

\item{y}{variable dependiente}

\item{x1}{variable independiente de nivel 1}

\item{x2}{variable independiente de nivel 2}

\item{peso_est}{peso de estudiante}

\item{peso_ie}{peso de escuela}

\item{idie}{id de escuela}

\item{r}{ruta y nombre del archivo Mplus. La funcion crea un archivo .dat y .out de Mplus.}
}
\value{
objeto mplus
}
\description{
Regresion multinivel con pesos. Una version mas fluida para usar las bondades de MplusAutomation. Debe estar instalado Mplus.
}
\examples{
\dontrun{
rr <- "D:/EM22/02-reportes-scripts/03-cruce-con-rendimiento/2-modelos-ffaa/mods/mod"
m00 <- reg_mplus(bd1, y = "M500_L", x1 = "ise"
peso_est = "Peso_mate", peso_ie = "peso_ie",
idie = "cod_mod7",
r = rr)
}

}
