% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpluscoef.R
\name{mpluscoef}
\alias{mpluscoef}
\title{Parametros estimados de mplus_hlm}
\usage{
mpluscoef(mplus_hlm_salida)
}
\arguments{
\item{mplus_hlm_salida}{objeto de mplus_hlm}
}
\value{
Devuelve parametros estimados de mplus_hlm
}
\description{
Parametros estimados de mplus_hlm
}
\examples{
\dontrun{
mpluscoef(m1)
}

}
