% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpluscoef.R
\name{mplusicc}
\alias{mplusicc}
\title{ICC estimado de mplus_hlm}
\usage{
mplusicc(mplus_hlm_salida)
}
\arguments{
\item{mplus_hlm_salida}{objeto de mplus_hlm}
}
\value{
Devuelve ICC estimado de mplus_hlm
}
\description{
ICC estimado de mplus_hlm
}
\examples{
\dontrun{
mplusicc(m1)
}



}
