% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_recursivo.R
\name{pca_recursivo}
\alias{pca_recursivo}
\title{Realiza analisis de componentes principales buscando una solucion razonable}
\usage{
pca_recursivo(data, recursivo = TRUE, puntajes = TRUE)
}
\arguments{
\item{data}{dataframe}

\item{recursivo}{TRUE si se quiere recursivo}

\item{puntajes}{TRUE en caso se requieran los puntajes}
}
\value{
lista con la varianza explicada, cargas y puntajes del primer componente
}
\description{
Si recursivo = TRUE, ejecuta un PCA hasta encontrar por lo menos una varianza explicada de .50
}
\examples{
\dontrun{
pca_recursivo(iris[1:4], recursivo = TRUE)
}

}
