% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_cfa_lavaan.R
\name{reporte_cfa_lavaan}
\alias{reporte_cfa_lavaan}
\title{Acomoda información del CFA para el reporte}
\usage{
reporte_cfa_lavaan(model_cfa_lavaan, puntajes = TRUE)
}
\arguments{
\item{model_cfa_lavaan}{objeto cfa de lavaan}

\item{puntajes}{si quieres puntajes}
}
\value{
lista con indicadores de ajuste, cargas factoriales y confiabilidad
}
\description{
Toma un objeto cfa del paquete lavaan y acomoda la información para el reporte psicometrico
}
\examples{
\dontrun{
HS.model <- ' visual  =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed   =~ x7 + x8 + x9 '
fit <- lavaan::cfa(HS.model, data = lavaan::HolzingerSwineford1939
reporte_cfa_lavaan(fit, puntajes = FALSE)
}

}
