% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_pca.R
\name{reporte_pca}
\alias{reporte_pca}
\title{Analisis de componentes principales (PCA)}
\usage{
reporte_pca(x, corr = NULL, puntajes = TRUE)
}
\arguments{
\item{x}{dataframe}

\item{corr}{con "poly" realiza correlacion policorica}

\item{puntajes}{si es que se quiere puntajes}
}
\value{
lista con la varianza explicada, cargas, alpha de cronbach y puntajes del primer componente
}
\description{
Realiza un analisis de componentes principales y devuelve la varianza explicada, las cargas y puntajes del primer componente.
}
\examples{

reporte_pca(iris[1:4])

}
