% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablaf.R
\name{tablaf}
\alias{tablaf}
\title{Tabla de frecuencias (Frequency table)}
\usage{
tablaf(data, x, peso = NULL, na = NULL, compara = NULL)
}
\arguments{
\item{data}{data frame}

\item{x}{variable categorica (categorical variable)}

\item{peso}{en caso hayan pesos (if there are weights)}

\item{na}{en caso se consideren valores perdidos}

\item{compara}{util para comparar los porcentajes segun el grupo}
}
\value{
data frame con la cantidad de casos (n) y porcentaje (prop)
}
\description{
Lo mismo que table() pero devuelve un data frame y acepta pesos (Same as table() but returns a dataframe and accepts weights)
}
\examples{
tablaf(mtcars, "gear")
}
