% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablaf_col.R
\name{tablaf_col}
\alias{tablaf_col}
\title{Tabla de frecuencias con varias columnas (Frequency table with many columns)}
\usage{
tablaf_col(data, nomvar, peso = NULL, starts = NULL, na = NULL)
}
\arguments{
\item{data}{data frame}

\item{nomvar}{vector con el nombre de la variable categorica (categorical variable)}

\item{peso}{en caso hayan pesos (if there are weights)}

\item{starts}{TRUE si se quiere aplicar sobre columnas que comparten los mismos caracteres al inicio}

\item{na}{en caso se consideren valores perdidos ()}
}
\value{
data frame con la cantidad de casos (n) y porcentaje (prop)
}
\description{
Aplica \code{\link[=tablaf]{tablaf()}} a varias columnas
}
\examples{
tablaf_col(iris, "Sepal", starts = TRUE)

}
